* Refactored: Updated and optimized the plugin codebase to align with the latest WordPress coding standards, ensuring improved compatibility and performance. * Fixed: An issue with the status icon not displaying correctly during the Re-establish Connection process, enhancing the user's ability to monitor connection status effectively. * Fixed: The absence of "Next Run" information in the Cron Schedules table, ensuring complete and accurate scheduling data is visible. * Fixed: The display and functionality of the Clients table length menu, ensuring a smoother user interaction when adjusting table views. * Fixed: Issues with overlapping tooltip elements in the "Add New" dropdown menu items, improving clarity and usability of the navigation interface.

+++ class/class-mainwp-system-cron-jobs.php
@@ -593,7 +593,7 @@ class MainWP_System_Cron_Jobs {
 			MainWP_Utility::update_option( 'mainwp_updatescheck_last', $today_m_y );
 
 			// send http check notification.
-			if ( 1 === get_option( 'mainwp_check_http_response', 0 ) ) {
+			if ( 1 === (int) get_option( 'mainwp_check_http_response', 0 ) ) {
 				$this->start_notification_http_check( $plain_text );
 			}
